% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/square_val.R
\name{add_var}
\alias{add_var}
\title{A sum value}
\usage{
add_var(x, y)
}
\arguments{
\item{x}{A numeric value}

\item{y}{A numeric value}
}
\value{
A number
}
\description{
A sum value
}
\examples{
add_var(100, 200)
}
