% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/square_val.R
\name{square_val}
\alias{square_val}
\title{A square value}
\usage{
square_val(x)
}
\arguments{
\item{x}{A numeric value}
}
\value{
A number
}
\description{
A square value
}
\examples{
square_val(2)
square_val(c(1:10))
}
